% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_Cluster_parm.R
\docType{data}
\name{example_Cluster_parm}
\alias{example_Cluster_parm}
\title{Example cluster item parameter tables}
\format{
An object of class data.frame with columns below\cr\cr
a: slope parameters for each assertion\cr\cr
b: difficulty parameters for each assertion\cr\cr
cluster_var: cluster variance parameters for each assertion\cr\cr
position: cluster (relative) position in the test\cr\cr
ItemID: Item IDs for each assertion\cr\cr
AssertionID: Assertion IDs for each assertion\cr\cr
***\code{Cluster_parm} columns must follow the above order
}
\usage{
data(example_Cluster_parm)
}
\description{
Example cluster item parameter dataframe for the \code{Cluster_parm}
}
\keyword{datasets}
