% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmult.R
\docType{data}
\name{func}
\alias{func}
\title{Using Rcpp to create a specialized matrix multiplication function used in person_fit()}
\format{
An object of class \code{character} of length 1.
}
\usage{
func
}
\value{
a function named \code{mmult()}
}
\description{
Create specialized matrix multiplication function used in person_fit() while calculating the varaince of
loglikelihodd for cluster items
}
\author{
Zhongtian Lin lzt713@gmail.com
}
\keyword{datasets}
