% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_fit.R
\name{item_fit}
\alias{item_fit}
\title{Item fit statistics and fit plot}
\usage{
item_fit(
  SA_dat = NULL,
  Cluster_dat = NULL,
  SA_parm = NULL,
  Cluster_parm = NULL,
  Dv = 1,
  n.nodes = 21,
  est_theta = NULL,
  what = c("X2", "G2", "phi", "fitplot"),
  sd_overall = 1,
  missing_code = NULL,
  fitplot_arg = list(n_bin = 25, binsize = 50, collapsed = TRUE, LOT = -3, HOT = 3, BY
    = 0.01)
)
}
\arguments{
\item{SA_dat}{For one student, a vector of response to standalone items.
For more than one student, a matrix or dataframe of response to standalone items. One assertion per column.
Column order must match row order in \code{SA_parm}. Use NA for missing responses}

\item{Cluster_dat}{For one student, a vector of response to cluster items.
For more than one student, a matrix or dataframe of response to cluster items. One assertion per column.
Column order must match row order in \code{Cluster_parm}. Use NA for missing responses.}

\item{SA_parm}{a matrix or dataframe of item parameters for standalone items, where columns are
a (slope), b1, b2, ..., b_k (difficulty or step difficulty), g (guessing), ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_SA_parm} for an example. Use \code{?example_SA_parm} for detailed column descriptions}

\item{Cluster_parm}{a matrix or dataframe of item parameters for cluster items, where columns are
a (slope), b (difficulty), cluster variance, cluster position, ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_Cluster_parm} for an example. Use \code{?example_Cluster_parm} for detailed column descriptions}

\item{Dv}{scaling factor for IRT model (usually 1 or 1.7)}

\item{n.nodes}{number of nodes used when integrating out the nuisance dimension}

\item{est_theta}{a vector of estimated theta. The length of the vector should be equal to the number of students}

\item{what}{a character vector specifying what to compute. Possible values are "X2", "G2", "phi", and "fitplot".
See \strong{Value} section for details.}

\item{sd_overall}{variance of the (overall) theta}

\item{missing_code}{code for missing response}

\item{fitplot_arg}{a named list of the arguments that can be specified when creating fit plots. Possible names are \cr
\code{binsize}, a numerical for the minimum size of bins when grouping students by their theta values
\code{collapsed}, a logical indicating whether to collapse the bins when
\code{LOT}, \code{HOT}, \code{BY}: when plotting expected ICC, a sequence of hypothetical theta is needed.
these arguments specify the lower bound, upper bound and the step when creating the sequence}
}
\value{
A list of results specified in \code{what}. Possible results in the list are: \cr
 \code{X2} A list of 2. The first and second element are lower triangle matrices of Chi-square values and
 standardized Chi-square values, respectively, for every pair of assertions \cr
 \code{G2} A lower triangle matrix of G2 values for every pair of assertions \cr
 \code{phi} A lower triangle matrix of the phi coefficient difference between
 the observed and expected responses, for every pair of assertions \cr
 \code{fitplot} Observed and expected ICC plot for every assertion
}
\description{
Computing item fit statistics including Pearson's Chi-squared test,
G2 logliklihood-ratio test, Phi coefficient difference, as well as fit plots
Model currently supported are 1-3 PL, and Rasch Testlet, and a mix of these models.
Use \code{what} argument to specify one or more results to return. See \strong{Value} section for details.
}
\note{
If the test does not have SA items or Cluster items, use default (NULL) for the corresponding data and parameter arguments. \cr\cr
Rasch SA items can be treated as clusters. To do so, store SA item parameters in the \code{Cluster_parm} argument with 0 variances.
}
\section{Dependencies}{

\describe{
 \item{statmod, dplyr}
}
}

\examples{
\dontrun{
data("example_SA_parm")
data("example_Cluster_parm")
sigma <- diag(c(1, sqrt(unique(example_Cluster_parm$cluster_var))))
mu <- rep(0, nrow(sigma))
thetas <- MASS::mvrnorm(601,mu,sigma)
thetas[,1] <- seq(-3,3,0.01) #overall dimension theta values
itmDat <- sim_data(thetas = thetas, SA_parm = example_SA_parm, Cluster_parm = example_Cluster_parm)
SA_dat <- itmDat[,1:17]
Cluster_dat <- itmDat[,-1:-20]
example_SA_parm = example_SA_parm[1:17,] # only 3pl items
out_scoring <- list()
for (i in 1:nrow(itmDat)) {
  out_scoring[[i]] <- scoring(SA_dat[i,], Cluster_dat[i,], example_SA_parm, example_Cluster_parm, n.nodes = 11, SE=TRUE)
}
est_theta <- sapply(out_scoring, function(x) x$par)
out = item_fit(SA_dat=SA_dat,Cluster_dat=Cluster_dat,SA_parm=example_SA_parm,Cluster_parm=example_Cluster_parm,
               Dv=1,n.nodes=21,est_theta = est_theta, sd_overall=1,missing_code=NULL,
               fitplot_arg = list(n_bin=25, binsize=50, collapsed=TRUE, LOT=-3, HOT=3, BY=0.01))
}
}
\author{
Zhongtian Lin lzt713@gmail.com
}
