% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lord_wing.R
\name{lord_wing}
\alias{lord_wing}
\title{Lord-Wingersky algorithm for computing the marginal probability of the raw scores}
\usage{
lord_wing(
  cluster_var,
  a,
  b,
  theta,
  n.nodes = 21,
  return_additional = FALSE,
  Dv = 1
)
}
\arguments{
\item{cluster_var}{a vector of  of length J with repeated values of the cluster
variance (e.g., rep(0.8, 9) for a 9-assertion cluster item).
Alternativaly, a scalar value of the cluster variance for the item.}

\item{a}{a vector of length J for the a (slope) parameters}

\item{b}{a vector of of length J for the b (difficulty) parameters}

\item{theta}{a vector of thetas of length N}

\item{n.nodes}{number of nodes used when integrating out the specific dimension}

\item{return_additional}{if TRUE, return a list containing the marginal probability as well
as some additional by-product of the function such as the conditional probability tables. See \strong{Value} section for details.}

\item{Dv}{scaling factor for IRT model (usually 1 or 1.7)}
}
\value{
When \code{return_additional = FALSE}, returns the marginal probability of raw scores, which is
a J+1 by N matrix, where J+1 is the number of possible raw scores \cr
When \code{return_additional = TRUE}, returns a list containing the marginal probability (\code{prk.marginal})
as well as \code{prob} and \code{qrob}, where \code{qrob} is a N by \code{n.nodes} by J array containing the conditional probability
of correct response for each theta at each node of the nuisance dimension for each assertion, and \code{qrob = 1 - prob}
}
\description{
For N persons (theta values), this function uses the algorithm by Lord & Wingersky (1984) to compute
the marginal probability of the raw scores for
a cluster item (with J dichotomously scored assertions) modeled by the Rasch testlet model, where the word "marginal" means
integrating out the nuisance dimension from the conditional likelihood of the cluster items.
}
\examples{
data(example_Cluster_parm)
# Compute on the first cluster, for 5 students
one_cluster_parm <- example_Cluster_parm[example_Cluster_parm$position == 1,]
rst <- lord_wing(one_cluster_parm$cluster_var , one_cluster_parm$a, one_cluster_parm$b,
 theta <- seq(-2,2,1), return_additional = TRUE)
}
\references{
Lord, F. M., & Wingersky, M. S. (1984). Comparison of IRT true-score and equipercentile observed-score "equatings".
\emph{Applied Psychological Measurement}, 8(4), 453-461.
}
\author{
Zhongtian Lin lzt713@gmail.com
}
