% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmult}
\alias{mmult}
\title{A specialized matrix multiplication}
\usage{
mmult(m, v)
}
\arguments{
\item{m, }{a numeric matrix}

\item{v, }{a numeric vector}
}
\description{
A specialized matrix multiplication function used in person_fit() while calculating the variance of
loglikelihodd for cluster items
}
\author{
Zhongtian Lin lzt713@gmail.com
}
