% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observed Data Marginal Likelihood Function.R
\name{obs.data.loglik}
\alias{obs.data.loglik}
\title{Marginal probability of the observed data}
\usage{
obs.data.loglik(
  theta,
  SA_dat = NULL,
  Cluster_dat = NULL,
  SA_parm = NULL,
  Cluster_parm = NULL,
  Dv = 1,
  n.nodes = 21,
  return_additional = F,
  missing_as_incorrect = F
)
}
\arguments{
\item{theta}{a scalar or a vector of student ability}

\item{SA_dat}{For one student: a vector of responses to standalone items \cr\cr
For two or more students: a matrix or dataframe of responses to standalone items. One assertion per column. Column order must match row order in SA_parm \cr\cr
Use NA for missing reponses}

\item{Cluster_dat}{For one student: a vector of responses to cluster items \cr\cr
For two or more students: a matrix or dataframe of responses to cluster items. One assertion per column. Column order must match row order in Cluster_parm \cr\cr
Use NA for missing reponses}

\item{SA_parm}{a matrix or dataframe of a,b,g parameters (column must be in this order), ItemID, and Assertion_ID for SA items}

\item{Cluster_parm}{a matrix or dataframe with columns of a, b and variance parameters for each assertion, as well as a column of cluster position, a column of cluster ItemID, and a column of Assertion_ID for Cluster items}

\item{Dv}{scaling factor for IRT model [use 1 or 1.7]}

\item{n.nodes}{number of nodes used when integrating out the specific dimension}

\item{return_additional}{if TRUE, return a list of the loglik plus some additional by-product of the function such as: \cr\cr
probs.SA: \tabular{ll}{\tab probablity table of correct response for standalones}
probs.cluster \tabular{ll}{\tab (conditional) probablity table of correct response for clusters at each given node}}

\item{missing_as_incorrect}{by default, missings (NAs) are treated as missing; if TRUE, missings are treated as incorrect}
}
\value{
If \emph{return_additional} is FALSE, returns a dataframe with two columns: theta and marginalized data loglikelihood; \cr
If \emph{return_additional} is TRUE, returns the dataframe of loglikelihood plus additional tables \emph{probs.SA} and \emph{probs.cluster} in a list
}
\description{
Computes the marginal probability of the observed data
}
\note{
If the test does not have SA items or Cluster items, use default (NULL) for the corresponding data and parameter arguments \cr\cr
SA items can be treated as clusters. To do so, store SA item parameters in the "Cluster_parm" argument with 0 variances, and store all student responses in "Cluster_dat"
}
\examples{

}
\author{
Zhongtian Lin lzt713@gmail.com
}
