% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoring}
\alias{scoring}
\title{IRT person scoring}
\usage{
scoring(
  SA_dat = NULL,
  Cluster_dat = NULL,
  SA_parm = NULL,
  Cluster_parm = NULL,
  Dv = 1,
  n.nodes = 21,
  censor = c(-4, 4),
  correction_val = 0.5,
  SE = FALSE
)
}
\arguments{
\item{SA_dat}{For one student, a vector or row matrix/dataframe of response to standalone items.
Responses must follow the row order in \code{SA_parm}. Use NA for missing responses.}

\item{Cluster_dat}{For one student, a vector or row matrix/dataframe of response to cluster items.
Responses must follow the row order in \code{Cluster_parm}. Use NA for missing responses.}

\item{SA_parm}{a matrix or dataframe of item parameters for standalone items, where columns are
a (slope), b1, b2, ..., b_k (difficulty or step difficulty), g (guessing), ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_SA_parm} for an example. Use \code{?example_SA_parm} for detailed column descriptions}

\item{Cluster_parm}{a matrix or dataframe of item parameters for cluster items, where columns are
a (slope), b (difficulty), cluster variance, cluster position, ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_Cluster_parm} for an example. Use \code{?example_Cluster_parm} for detailed column descriptions}

\item{Dv}{scaling factor for IRT model (usually 1 or 1.7)}

\item{n.nodes}{number of nodes used when integrating out the nuisance dimension}

\item{censor}{when there's perfect or all wrong score, this is a 2-element vector of upper and lower limit of \cr
1) starting value in point estimate optimization \cr
2) pseudo parameter estimate used in SE computation to avoid extremely large SE estimate}

\item{correction_val}{a value to add or subtract when there's perfect or all wrong score to avoid extremely large theta estimate}

\item{SE}{if TRUE, returns standard error}
}
\value{
a list of scoring results, where the first element is the estimated theta.
}
\description{
Compute IRT latent scores (theta estimates). Model currently supported are 1-3 PL, GPCM, Rasch Testlet,
and a mix of these models. see \code{?MIRTutils} for model details. \cr
Scoring method currently supported are: \cr
 MMLE, marginal maximum likelihood estimation. When no Rasch testlet model item is involved,
 it reduces to regular MLE.
}
\note{
If the test does not have SA items or Cluster items, use default (NULL) for the corresponding data and parameter arguments \cr\cr
Rasch SA items can be treated as clusters. To do so, store SA item parameters in the \code{Cluster_parm} argument with 0
variances, and store all student responses in \code{Cluster_dat} \cr\cr
This function should be run for one student at a time. Use parallel processing for higher processing speed.\cr\cr
DONOTUSE: \cr
SA_dat and Cluster_dat can be multi-row matrix/dataframe where it is one person per row. In that case, it simultaneously solves
all thetas but it assumes thetas are not independent. So it shouldn't be used for scoring purpose.
}
\section{Dependencies}{

\describe{
 \item{statmod}
}
}

\examples{
data(example_SA_parm)
data(example_Cluster_parm)
sigma <- diag(c(1, sqrt(unique(example_Cluster_parm$cluster_var))))
mu <- rep(0, nrow(sigma))
thetas <- MASS::mvrnorm(7,mu,sigma)
thetas[,1] <- seq(-3,3,1) #overall dimension theta values
itmDat <- sim_data(thetas = thetas, SA_parm = example_SA_parm, Cluster_parm = example_Cluster_parm)
SA_dat <- itmDat[,1:20]
Cluster_dat <- itmDat[,-1:-20]
# Scoring for the first student
rst <- scoring(SA_dat[1,], Cluster_dat[1,], example_SA_parm, example_Cluster_parm, n.nodes = 11, SE=TRUE)
rst$par  # estimated theta
rst$SE  # estimated standard error
}
\author{
Zhongtian Lin lzt713@gmail.com
}
