% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{utility}
\alias{utility}
\title{A utility function computing some commonly used values in IRT}
\usage{
utility(
  theta,
  SA_parm = NULL,
  Cluster_parm = NULL,
  Dv = 1,
  n.nodes = 50,
  what = c("prob", "escore", "iteminfo", "mprob")
)
}
\arguments{
\item{theta}{a scalar or a vector of examinee ability}

\item{SA_parm}{a matrix or dataframe of item parameters for standalone items, where columns are
a (slope), b1, b2, ..., b_k (difficulty or step difficulty), g (guessing), ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_SA_parm} for an example. Use \code{?example_SA_parm} for detailed column descriptions}

\item{Cluster_parm}{a matrix or dataframe of item parameters for cluster items, where columns are
a (slope), b (difficulty), cluster variance, cluster position, ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_Cluster_parm} for an example. Use \code{?example_Cluster_parm} for detailed column descriptions}

\item{Dv}{scaling factor for IRT model (usually 1 or 1.7)}

\item{n.nodes}{number of nodes used when integrating out the nuisance dimension}

\item{what}{a character vector specifying what to compute. Possible values are "prob", "escore", "iteminfo", and "mprob".
See \strong{Value} section for details.}
}
\value{
A list of results specified in \code{what}: \cr
\itemize{
\item \code{prob}: probability of correct response. When Rasch testlet model is involved, this is conditional probability given both overall and each node of the nuisance dimension
\item \code{escore}: expected item score
\item \code{iteminfo}: item information. When Rasch testlet model is involved, this is item info based on marginal prob of raw scores calculated using Lord-Wingerskey algorithm
\item \code{mprob}: When Rasch testlet model is involved, this is the marginal probability of correct response for each assertion. For unidimensional models, this is the same as prob
}
}
\description{
Compute some commonly used values in IRT such as
probability of correct response, expected item scores, and item information.
Use \code{what} argument to specify one or more results to return. See \strong{Value} section for details. \cr\cr
Run '?MIRTutils-package' for more details, such as the context of the current package and models supported.
}
\note{
If the test does not have SA items or Cluster items, use default (NULL) for the corresponding data and parameter arguments. \cr\cr
}
\examples{
data(example_SA_parm)
data(example_Cluster_parm)
theta <- seq(-3,3,1)
rst <- utility(theta, example_SA_parm, example_Cluster_parm, n.nodes = 11)
}
\author{
Zhongtian Lin \href{mailto:lzt713@gmail.com}{lzt713@gmail.com}
}
