% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loglik.R
\name{data_loglik}
\alias{data_loglik}
\title{Marginal loglikelihood of the observed data}
\usage{
data_loglik(
  theta,
  SA_dat = NULL,
  Cluster_dat = NULL,
  SA_parm = NULL,
  Cluster_parm = NULL,
  Dv = 1,
  n.nodes = 21,
  return_additional = TRUE,
  missing_as_incorrect = F
)
}
\arguments{
\item{theta}{a scalar or a vector of examinee ability}

\item{SA_dat}{For one examinee, a vector of response to standalone items.
For more than one examinee, a matrix or dataframe of response to standalone items. One assertion per column.
Column order must match row order in \code{SA_parm}. Use NA for missing responses}

\item{Cluster_dat}{For one examinee, a vector of response to cluster items.
For more than one examinee, a matrix or dataframe of response to cluster items. One assertion per column.
Column order must match row order in \code{Cluster_parm}. Use NA for missing responses.}

\item{SA_parm}{a matrix or dataframe of item parameters for standalone items, where columns are
a (slope), b1, b2, ..., b_k (difficulty or step difficulty), g (guessing), ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_SA_parm} for an example. Use \code{?example_SA_parm} for detailed column descriptions}

\item{Cluster_parm}{a matrix or dataframe of item parameters for cluster items, where columns are
a (slope), b (difficulty), cluster variance, cluster position, ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_Cluster_parm} for an example. Use \code{?example_Cluster_parm} for detailed column descriptions}

\item{Dv}{scaling factor for IRT model (usually 1 or 1.7)}

\item{n.nodes}{number of nodes used when integrating out the nuisance dimension}

\item{return_additional}{if TRUE, returns data loglikelihood with some additional by-product of the function in a list. See \strong{Value} section for details\cr}

\item{missing_as_incorrect}{by default, missings (NAs) are treated as missing; if TRUE, missings are treated as incorrect}
}
\value{
If \code{return_additional} is \code{FALSE}, returns a dataframe with two columns: theta and marginalized data loglikelihood. \cr\cr
If \code{return_additional} is \code{TRUE}, returns the dataframe of loglikelihood with following additional tables in a list
\itemize{
\item \code{probs.SA}: probability of correct response for standalone items \cr
\item \code{probs.cluster}: (conditional) probability table of correct response for clusters at each given nodes \cr
\item \code{parms}: parameter tables in a list
}
}
\description{
Compute observed data marginal loglikihood.
\itemize{
\item When there are only standalone items, this computes the regular loglik of the observed data.
\item When there are cluster items, this computes the marginal loglik where the word marginal means
integrating out the nuisance dimension from the conditional likelihood of the cluster items. \cr\cr
Use '?MIRTutils-package' for more details, such as the context of the current package and models supported.
}
}
\note{
If the test does not have SA items or Cluster items, use default (NULL)
for the corresponding data and parameter arguments \cr\cr
}
\examples{
data(example_SA_parm)
data(example_Cluster_parm)
sigma <- diag(c(1, sqrt(unique(example_Cluster_parm$cluster_var))))
mu <- rep(0, nrow(sigma))
thetas <- MASS::mvrnorm(7,mu,sigma)
thetas[,1] <- seq(-3,3,1) #overall dimension theta values
itmDat <- sim_data(thetas = thetas, SA_parm = example_SA_parm, Cluster_parm = example_Cluster_parm)
SA_dat <- itmDat[,1:20]
Cluster_dat <- itmDat[,-1:-20]
rst <- data_loglik(thetas[,1], SA_dat, Cluster_dat, example_SA_parm, example_Cluster_parm, n.nodes = 11, return_additional = TRUE)
rst$loglik
rst$prob.SA
length(rst$probs.cluster) # a list conditional probabilities. The length of the list = number of clusters
}
\author{
Zhongtian Lin \href{mailto:lzt713@gmail.com}{lzt713@gmail.com}
}
