% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lord_wing.R
\name{lord_wing}
\alias{lord_wing}
\title{Lord-Wingersky algorithm for computing the marginal probability of the raw scores}
\usage{
lord_wing(
  cluster_var,
  a,
  b,
  theta,
  n.nodes = 21,
  return_additional = FALSE,
  Dv = 1
)
}
\arguments{
\item{cluster_var}{a vector of  of length J with repeated values of the cluster
variance (e.g., rep(0.8, 9) for a 9-assertion cluster item).
Alternatively, a scalar value of the cluster variance for the item.}

\item{a}{a vector of length J for the a (slope) parameters}

\item{b}{a vector of of length J for the b (difficulty) parameters}

\item{theta}{a vector of length N for the thetas}

\item{n.nodes}{number of nodes used when integrating out the specific dimension}

\item{return_additional}{if TRUE, return a list containing the marginal probability as well
as some additional by-product of the function such as the conditional probability tables. See \strong{Value} section for details.}

\item{Dv}{scaling factor for IRT model (usually 1 or 1.7)}
}
\value{
When \code{return_additional = FALSE}, returns the marginal probability of raw scores, which is
a J+1 by N matrix, where J+1 is the number of possible raw scores \cr\cr
When \code{return_additional = TRUE}, returns a list containing
\itemize{
\item \code{prk}: the marginal probability
\item \code{prob}: N by \code{n.nodes} by J array containing the conditional probability
of correct response for each theta at each node of the nuisance dimension for each assertion
\item \code{qrob}: \code{1 - prob}
\item \code{nodes} and \code{whts}: nodes and weights used in the calculation.
}
}
\description{
For N persons (theta values), this function makes use of the recursive algorithm
by Lord & Wingersky (1984) and its extensions (Lin et.al, 2024) to compute
the marginal probability of the raw scores for a cluster
item (with J dichotomously scored assertions) modeled by the Rasch testlet model.
The word "marginal" here means integrating out the nuisance dimension from the
conditional likelihood of the cluster items.
}
\examples{
data(example_Cluster_parm)
# Compute on the first cluster, for 5 students
one_cluster_parm <- example_Cluster_parm[example_Cluster_parm$position == 1,]
rst <- lord_wing(one_cluster_parm$cluster_var , one_cluster_parm$a, one_cluster_parm$b,
theta <- seq(-2,2,1), return_additional = TRUE)
}
\references{
Lord, F. M., & Wingersky, M. S. (1984). Comparison of IRT true-score and equipercentile observed-score "equatings".
\emph{Applied Psychological Measurement}, 8(4), 453-461. \cr\cr
Lin, Z., Jiang, T., Rijmen, F. et al. (2024). Asymptotically Correct Person Fit z-Statistics For the Rasch Testlet Model.
\emph{Psychometrika}, https://doi.org/10.1007/s11336-024-09997-y.
}
\author{
Zhongtian Lin \href{mailto:lzt713@gmail.com}{lzt713@gmail.com}
}
