% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_fit.R
\name{person_fit}
\alias{person_fit}
\title{Person fit index computation}
\usage{
person_fit(
  theta,
  SA_dat = NULL,
  Cluster_dat = NULL,
  SA_parm = NULL,
  Cluster_parm = NULL,
  Dv = 1,
  Alpha = 0.05,
  n.nodes = 21
)
}
\arguments{
\item{theta}{a scalar or a vector of examinee ability}

\item{SA_dat}{For one examinee, a vector of response to standalone items.
For more than one examinee, a matrix or dataframe of response to standalone items. One assertion per column.
Column order must match row order in \code{SA_parm}. Use NA for missing responses}

\item{Cluster_dat}{For one examinee, a vector of response to cluster items.
For more than one examinee, a matrix or dataframe of response to cluster items. One assertion per column.
Column order must match row order in \code{Cluster_parm}. Use NA for missing responses.}

\item{SA_parm}{a matrix or dataframe of item parameters for standalone items, where columns are
a (slope), b1, b2, ..., b_k (difficulty or step difficulty), g (guessing), ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_SA_parm} for an example. Use \code{?example_SA_parm} for detailed column descriptions}

\item{Cluster_parm}{a matrix or dataframe of item parameters for cluster items, where columns are
a (slope), b (difficulty), cluster variance, cluster position, ItemID, and AssertionID.
Columns must follow the above order.
See \code{example_Cluster_parm} for an example. Use \code{?example_Cluster_parm} for detailed column descriptions}

\item{Dv}{scaling factor for IRT model (usually 1 or 1.7)}

\item{Alpha}{a numeric vector for one or more nominal type I error rates for flagging aberrant responses.
For example, \code{c(0.01,0.05)}}

\item{n.nodes}{number of nodes used when integrating out the nuisance dimension}
}
\value{
a one-row dataframe with the following columns: \cr
\itemize{
\item \code{theta}: theta score user supplied \cr
\item \code{raw}: raw score \cr
\item \code{ll}: observed data loglikelihood based on MMLE \cr
\item \code{exp_ll}: expected value of the data loglikelihood \cr
\item \code{var_ll_org}: variance of the data loglikelihood before correction \cr
\item \code{correction}: correction of the variance of the data loglikelihood \cr
\item \code{var_ll}: variance of the data loglikelihood after correction \cr
\item \code{lz}: lz* or lzt* person fit index \cr
\item \code{flag_[Alpha]}: one or more columns of flags for aberrant responses based on lz and the
type I error rates specified in \code{Alpha}
}
}
\description{
Calculates person fit statistics. Models currently supported are 1-3 PL, GPCM, Rasch Testlet, and a mix of these models.
\itemize{
\item For traditional test (i.e., standalone items only), this is lz* described in Snijders (2001).
\item For Cluster items, this is lzt* described in Lin et.al (2024), which extends the lz* to the Rasch testlet
model when theta is estimated by MMLE, .
}
}
\note{
This function should be run for one examinee at a time. Use parallel processing for higher processing speed. \cr\cr
If the test does not have SA items or Cluster items, use default (NULL) for the corresponding data and parameter arguments \cr\cr
If one prefer, Rasch SA items can be treated as clusters. To do so, store SA item parameters in the \code{Cluster_parm}
argument with 0 variances, and store all examinee responses in \code{Cluster_dat} \cr\cr
}
\examples{
data(example_SA_parm)
data(example_Cluster_parm)
sigma <- diag(c(1, sqrt(unique(example_Cluster_parm$cluster_var))))
mu <- rep(0, nrow(sigma))
thetas <- MASS::mvrnorm(7,mu,sigma)
thetas[,1] <- seq(-3,3,1) #overall dimension theta values
itmDat <- sim_data(thetas = thetas, SA_parm = example_SA_parm, Cluster_parm = example_Cluster_parm)
SA_dat <- itmDat[,1:20]
Cluster_dat <- itmDat[,-1:-20]
# Scoring for the first examinee
out_scoring <- scoring(SA_dat[1,], Cluster_dat[1,], example_SA_parm, example_Cluster_parm, n.nodes = 11, SE=TRUE)
est_theta <- out_scoring$par
# person fit for the first examinee
rst <- person_fit(est_theta, SA_dat[1,], Cluster_dat[1,], example_SA_parm, example_Cluster_parm, Alpha=c(0.01,0.05))
}
\references{
Snijders, T. A. (2001). Asymptotic null distribution of person fit statistics with estimated person parameter.
\emph{Psychometrika}, 66(3), 331-342. \cr\cr
Lin, Z., Jiang, T., Rijmen, F. et al. (2024). Asymptotically Correct Person Fit z-Statistics For the Rasch Testlet Model.
\emph{Psychometrika}, https://doi.org/10.1007/s11336-024-09997-y.
}
\author{
Zhongtian Lin lzt713@gmail.com
}
